Functional logic within RegisterForm.jsx
